pyproject.toml
[tool.poetry]
name = "teamchallenge-1444"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "teamchallenge"}]

[tool.poetry.dependencies]
python = "^3.11,<4"
django = "^5.1.5"
djangorestframework = "^3.15.2"
django-jsoneditor = "^0.2.4"
drf-spectacular = "^0.28.0"
django-filter = "^24.3"
django-redis = "^5.4.0"
python-dotenv = "^1.0.1"
drf-nested-routers = "^0.94.1"
pillow = "^11.1.0"
djangorestframework-simplejwt = "^5.4.0"
qrcode = "^8.0"
django-environ = "^0.12.0"
pandas = "^2.2.3"
openpyxl = "^3.1.5"


[tool.poetry.group.deploy.dependencies]
gunicorn = "^23.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
